Use Fuel for data iteration.
This changes the format in which data is read: Dictionaries are now
assumed to be text files with tabs, the last column of which contains
the word (just one column is okay too, but this allows words to be
prefixed with word counts). It now expects you to pass a *list* of data
files instead of a single one.