EASGD parameters, shuffled batches, measuring time
For EASGD they generally set beta and use it calculate alpha as a
function of the communication period and the number of workers. Because
the controller is agnostic to the number of workers connected we must
pass it explicitly.

In order to plot the logs of multiple workers together we should log the
actual time (instead of the relative time).

The sorting of batches seems to lead to some strange training artifacts;
sometimes one of the workers gets significantly longer sentences than
the other for a while, which makes it look like it's doing really badly.
Moreover, the sentences sampled are always of the same length if
sample_freq divides the batch buffer. To guarantee that this doesn't
affect training too much I shuffled the batches. There are still some
artifacts (one worker gets shorter/longer sentnces for a while) it
reduces the noise a bit. In general it might be better to monitor
cost/target length instead of just the cost.